+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.12.0-beta.1 (Unreleased)
 - Added support for specifying an encryption scope.
+- Added support for the 2020-08-04 service version.
 
 ## 12.11.0 (2021-04-29)
 - GA release