+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.12.0-beta.1 (2021-05-11)
+## 12.12.0-beta.1 (2021-05-13)
 - Added support for specifying an encryption scope.
 - Added support for the 2020-08-04 service version.